reorganize herder_facade a bunch
