+++ src/herder_facade/facade.rs
@@ -1,9 +1,90 @@
+use super::client::MaybeHerder;
+use super::{HerderFacade, StartWriterError, WriterHandle};
+use crate::herder_daemon::ipc::{StatusMessage, WriterProcessConfig};
+use futures::StreamExt;
+use futures::stream::{self, BoxStream};
 use std::collections::HashMap;
 use std::hash::Hash;
 use std::sync::Arc;
-
-use futures::stream::{self, BoxStream};
 use tokio::sync::mpsc::{self, error::TryRecvError};
+use tracing::{debug, trace};
+
+/// The actual [HerderFacade] used by Caligula.
+pub struct HerderFacadeImpl {
+    event_demux: EventDemux<u64, StatusMessage>,
+    next_writer_id: u64,
+
+    standard_daemon: MaybeHerder,
+    escalated_daemon: MaybeHerder,
+}
+
+impl HerderFacadeImpl {
+    pub fn new(log_path: &str) -> Self {
+        let (writer_tx, writer_rx) = mpsc::unbounded_channel();
+
+        let cloned = writer_tx.clone();
+        let standard_daemon = MaybeHerder::new(
+            log_path.to_owned(),
+            false,
+            Box::new(move |e| {
+                cloned.send(e).unwrap();
+            }),
+        );
+
+        let escalated_daemon = MaybeHerder::new(
+            log_path.to_owned(),
+            true,
+            Box::new(move |e| {
+                writer_tx.send(e).unwrap();
+            }),
+        );
+
+        Self {
+            event_demux: EventDemux::new(writer_rx),
+            next_writer_id: 0,
+            standard_daemon,
+            escalated_daemon,
+        }
+    }
+}
+
+impl HerderFacade for HerderFacadeImpl {
+    fn start_writer(
+        &mut self,
+        args: &WriterProcessConfig,
+        escalated: bool,
+    ) -> impl Future<Output = Result<WriterHandle, StartWriterError>> {
+        async move {
+            let id = self.next_writer_id;
+            self.next_writer_id += 1;
+
+            let d = match escalated {
+                true => &mut self.escalated_daemon,
+                false => &mut self.standard_daemon,
+            };
+
+            d.start_writer(id, args).await?;
+
+            trace!("Reading results from child");
+            let mut event_rx = self.event_demux.select_key(id).unwrap();
+
+            let first_msg = event_rx.next().await;
+            debug!(?first_msg, "Read raw result from child");
+
+            let initial_info = match first_msg {
+                Some(StatusMessage::InitSuccess(i)) => Ok(i),
+                Some(StatusMessage::Error(t)) => Err(StartWriterError::Failed(Some(t))),
+                Some(other) => Err(StartWriterError::UnexpectedFirstStatus(other)),
+                None => Err(StartWriterError::UnexpectedDisconnect),
+            }?;
+
+            Ok(WriterHandle {
+                events: event_rx,
+                initial_info,
+            })
+        }
+    }
+}
 
 #[derive(Debug)]
 pub struct EventDemux<K: Hash + Eq + Send, T: Send> {
